benchmarks/prediction-pipeline/requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
numpy==1.16.3
pbr==5.2.0
pillow==3.1.2
protobuf==3.7.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.2
Werkzeug==0.15.2
